% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStats.R
\name{addStats}
\alias{addStats}
\title{addStats}
\usage{
addStats(o, randomize, effectPN, null, deltaL2, deltaL1, deltaG1, deltaG2)
}
\arguments{
\item{o}{Single mcmc replicate}

\item{randomize}{TRUE if allocation is vector has length > 1}

\item{effectPN}{See SeqSGPV}

\item{null}{See SeqSGPV}

\item{deltaL2}{See SeqSGPV}

\item{deltaL1}{See SeqSGPV}

\item{deltaG1}{See SeqSGPV}

\item{deltaG2}{See SeqSGPV}
}
\description{
Obtains bias, whether the interval rejects H0, coverage, sgpvROPE, and sgpvROME for a given interval.  addStats is called within SeqSGPV.
}
