% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amData.R
\name{amData}
\alias{amData}
\title{amData}
\usage{
amData(nreps, os, fork = TRUE, socket = TRUE, cores = detectCores(), ...)
}
\description{
Generate mcmc simulations of adaptive monitoring with parallel computing
}
