% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amDataGetMore.R
\name{amDataGetMore}
\alias{amDataGetMore}
\title{amDataGetMore}
\usage{
amDataGetMore(insufficients, existingDataList, fork = TRUE,
  socket = TRUE, cores = detectCores(), ...)
}
\description{
Generate mcmc simulations of adaptive monitoring with parallel computing
}
