% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amDataSingle.R
\name{amDataSingle}
\alias{amDataSingle}
\title{amDataSingle}
\usage{
amDataSingle(
  dataGeneration,
  dataGenArgs,
  effectGeneration,
  effectGenArgs,
  effectScale = "identity",
  randomize = FALSE,
  modelFit,
  modelFitArgs,
  monitoringIntervalLevel,
  existingData = NULL
)
}
\description{
Generate a single adaptive monitoring simulation
}
