% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amDataSingleGetMore.R
\name{amDataSingleGetMore}
\alias{amDataSingleGetMore}
\title{amDataSingleGetMore}
\usage{
amDataSingleGetMore(
  existingDataList,
  iInsufficient,
  getMore,
  monitoringIntervalLevel,
  dataGeneration,
  dataGenArgs,
  effectGeneration,
  effectGenArgs,
  effectScale,
  randomize,
  pointNull,
  deltaL2,
  deltaL1,
  deltaG1,
  deltaG2,
  modelFit
)
}
\description{
Generate additional observations for single mcmc simulation
}
