% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfDataReduction.R
\name{ecdfDataReduction}
\alias{ecdfDataReduction}
\title{ecdfDataReduction.R}
\usage{
ecdfDataReduction(x, n = 200)
}
\arguments{
\item{x}{observations}

\item{n}{desired number of reduced, ordered observations}
}
\value{
ecdf after reducing (if applicable) the number of observations
}
\description{
Order and reduce x observations before obtaining ecdf.  Helps save storage space when saving ecdf function to object and when drawing upon many observations.  ecdfDataReduction is called within SeqSGPV.
}
