% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCI.R
\name{fitCI}
\alias{fitCI}
\alias{lmCI}
\alias{lrCI}
\title{Functions for obtaining Confidence Intervals under a linear model and under logistic regression}
\usage{
lmCI(y, trt, look, miLevel)

lrCI(y, trt, look, miLevel)
}
\arguments{
\item{y}{outcomes}

\item{trt}{treatment assignments for each patient}

\item{look}{current number of patients accrued}

\item{miLevel}{alpha in traditional (1-alpha) confidence interval used in monitoring intervals}
}
\value{
est, lowwer bound, and upper bound
}
\description{
Functions for obtaining Confidence Intervals under a linear model and under logistic regression
}
