% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedDesignEffects.R
\name{fixedDesignEffects}
\alias{fixedDesignEffects}
\title{Function to assess operating characteristics under a range of shifted treatment effects}
\usage{
fixedDesignEffects(o, shift, printProgress = TRUE)
}
\arguments{
\item{o}{Object from SeqSGPV}

\item{shift}{Alternative effects to assess operating characteristics}

\item{printProgress}{TRUE to print progress}
}
\description{
Function to assess operating characteristics under a range of shifted treatment effects
}
