% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCI.R
\name{lrCI}
\alias{lrCI}
\title{lrCI}
\usage{
lrCI(look, modelFitArgs)
}
\arguments{
\item{look}{Current number of observations accrued}

\item{modelFitArgs}{Arguments specific to the interval function.  lmCI and lrCI require y (outcomes), XD (design matrix), and miLevel (monitoring interval level)  which are set in SeqSGPV. binomCI requires x (successes), n (observations), and any other inputs used in binom::binconf.}
}
\value{
est, lower bound, and upper bound
}
\description{
lrCI
}
