% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcMonitoringEnoughCheck.R
\name{mcmcMonitoringEnoughCheck}
\alias{mcmcMonitoringEnoughCheck}
\title{mcmcMontiroingEnoughCheck}
\usage{
mcmcMonitoringEnoughCheck(o, monitoringFrequency)
}
\arguments{
\item{o}{List of mcmc replicates}

\item{monitoringFrequency}{Matrix of monitoring frequency combinations}
}
\description{
Makes sure enough observations collected to ensure end of study when applying monitoring rules. mcmcMontiroingEnoughCheck is called within SeqSPGV.
}
