% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqSGPVrules.R
\name{SeqSGPVrules}
\alias{SeqSGPVrules}
\title{SeqSGPVrules}
\usage{
SeqSGPVrules(
  mcmcMonitoring,
  os,
  fork = TRUE,
  socket = TRUE,
  cores = parallel::detectCores(),
  monitoringFrequency,
  randomize
)
}
\arguments{
\item{mcmcMonitoring}{List of mcmc replicates}

\item{os}{See SeqSGPV}

\item{fork}{See SeqSGPV}

\item{socket}{See SeqSGPV}

\item{cores}{See SeqSGPV}

\item{monitoringFrequency}{Matrix of monitoring frequency combinations}

\item{randomize}{TRUE if length(allocations) > 1}
}
\description{
Applies monitoring frequencies to PRISM monitoring and returns the final observed observation. SeqSGPVrules is called within SeqSGPV.
}
