% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqSGPVrulesSingle.R
\name{SeqSGPVrulesSingle}
\alias{SeqSGPVrulesSingle}
\title{SeqSGPVrules}
\usage{
SeqSGPVrulesSingle(data, monitoringFrequency, randomize)
}
\arguments{
\item{data}{Single mcmc replicate}

\item{monitoringFrequency}{Matrix of monitoring frequency combinations}

\item{randomize}{TRUE if length(allocations) > 1}
}
\description{
Applies monitoring frequencies to PRISM monitoring and returns the final observed observation. SeqSGPVrules is called within SeqSGPV.
}
