% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpvAMrules.R
\name{sgpvAMrules}
\alias{sgpvAMrules}
\title{amSgpvRules}
\usage{
sgpvAMrules(
  mcmcMonitoring,
  os,
  fork = TRUE,
  socket = TRUE,
  cores = detectCores(),
  waitJ,
  lookS,
  affirmK,
  miLevel,
  getUnrestricted,
  maxN,
  lagN
)
}
\description{
Generate mcmc simulations of adaptive monitoring with parallel computing
}
