% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SeqSGPV.R
\name{summary.SeqSGPV}
\alias{summary.SeqSGPV}
\title{S@title summary.SeqSGPV}
\usage{
\method{summary}{SeqSGPV}(am, wait, steps, affirm, N, lag, effect, rd = 4)
}
\arguments{
\item{am}{SeqSGPV object.}

\item{wait}{Vector of possible wait times (W) before monitoring.}

\item{steps}{Vector of the number of observations (S) in between monitoring assessments.}

\item{affirm}{Vector of the number of observations required for affirming a stopping rule (A) in between raising an alert of stopping.}

\item{N}{Vector of maximum sample size. Can be set as Inf (indefinite).}

\item{lag}{Vector of the number of delayed outcomes between enrolling a single observation and observing its outcome.}

\item{effect}{Effect of interest.}

\item{rd}{Number of rounding decimals.}
}
\description{
Summary of the design-based operating characteristics conditioned on a single effect and a single combination of monitoring frequency parameters (wait, steps, affirm, lag, N).
}
