% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amDataGetMore.R
\name{amDataGetMore}
\alias{amDataGetMore}
\title{amDataGetMore}
\usage{
amDataGetMore(
  insufficients,
  existingDataList,
  os,
  fork = TRUE,
  socket = TRUE,
  cores = detectCores(),
  ...
)
}
\arguments{
\item{insufficients}{Column vector of indices corresponding to existingDataList that needs additional data}

\item{fork}{Fork clustering, works on POSIX systems (Mac, Linux, Unix, BSD) and not Windows.  Defaults to TRUE.}

\item{socket}{Socket clustering.  Defaults to TRUE yet only applies if FORK = FALSE.}

\item{...}{Inputs to amDataSingleGetMore}

\item{exisitingDataList}{List of generated mcmc data in each element.}
}
\description{
Generate additional data if previously generated data were insufficient to run to conclusion with unrestricted n.
}
