% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amDataSingleGetMore.R
\name{amDataSingleGetMore}
\alias{amDataSingleGetMore}
\title{amDataSingleGetMore}
\usage{
amDataSingleGetMore(existingDataList, iInsufficient, getMore,
  monitoringIntervalLevel, dataGeneration, dataGenArgs, effectGeneration,
  effectGenArgs, modelFit)
}
\description{
Generate additional observations for single mcmc simulation
}
