% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpvAM.R
\name{sgpvAM}
\alias{sgpvAM}
\title{Adaptive Monitoring with Second Generation p-Value}
\usage{
sgpvAM(mcmcData = NULL, nreps, waitWidths, dataGeneration = NULL,
  dataGenArgs, effectGeneration = NULL, effectGenArgs, modelFit,
  pointNull, deltaL2, deltaL1, deltaG1, deltaG2, lookSteps = 1,
  kSteps = 10, maxAlertSteps = 100, maxN = NULL, lagOutcomeN = 0,
  monitoringIntervalLevel = 0.05, printProgress = TRUE,
  outData = TRUE, getECDF = TRUE, cores = NULL, fork = TRUE,
  socket = TRUE)
}
\arguments{
\item{mcmcData}{Previously generated data.  Default (NULL) uses mcmc generation inputs to generate new or additional data.}

\item{nreps}{Number of mcmc replicates to generate}

\item{waitWidths}{Wait time, in terms of Confidence Interval width, before monitoring data.  Assuming the data generation standard deviation, waits until appropriate n to achieve confidence interval width.}

\item{dataGeneration}{Function (such as rnorm) to generate outcomes.}

\item{dataGenArgs}{Arguments for dataGeneration function.  This includes, in the least, 'n' observations to generate.  If 'n' is insufficient for unrestricted adaptive monitoring, addtional data will be generated.}

\item{effectGeneration}{Function (such as rnorm) or fixed value to generate intervention effect (theta).}

\item{effectGenArgs}{Arguments for effectGeneration function (if any)}

\item{modelFit}{A existing or user-defined function to obtain intervals.  Two existing functions are provided: 1) lmCI which obtains a confidence interval from a linear model and has class 'normal' indicating normal data and 2) lrCI obtains Wald Confidence Interval from logistic regression model and has class 'binomial' indicating binomial data.}

\item{pointNull}{Point null.}

\item{deltaL2}{Clinical guidepost less than and furthest from point null.}

\item{deltaL1}{Clinical guidepost less than and closest to point null.}

\item{deltaG1}{Clinical guidepost greater than and closest to point null.}

\item{deltaG2}{Clinical guidepost greater than and furthest from point null.}

\item{lookSteps}{The frequency data are observed (defaults to 1 -- fully sequential).}

\item{kSteps}{Affirmation steps to consider range from 0 to maxAlertSteps by kSteps.}

\item{maxAlertSteps}{Maximum number of steps before affirming an alert.}

\item{maxN}{Total enrolled patients equals maxN observed patients plus lagOutcomeN.}

\item{lagOutcomeN}{Total enrolled patients equals MaxN observed patients plus lagOutcomeN.  lagOutcomeN are number of observations enrolled but awaiting to observe outcome.}

\item{monitoringIntervalLevel}{Traditional (1-alpha) used in monitoring intervals.}

\item{printProgress}{Prints when adding more data for mcmc replicates to have sufficient observations to monitor until a conclusion.  Defaults to TRUE.}

\item{outData}{Returns the mcmc generated data.  This can result in an out object with large memory.  Yet, with location shift data, can be re-used to obtain operating characteristics of shifted effects.}

\item{getECDF}{Returns the ECDF of sample size and bias for each wait width and number of steps before affirming end of study.}

\item{cores}{Number of cores used in parallel computing.  The default (NULL) does not run on parallel cores.}

\item{fork}{Fork clustering, works on POSIX systems (Mac, Linux, Unix, BSD) and not Windows.  Defaults to TRUE.}

\item{socket}{Socket clustering.  Defaults to TRUE yet only applies if FORK = FALSE.}
}
\value{
\enumerate{
            \item [mcmcMonitoring] The monitored generated / pre-generated datasets
            \item }
mcmcMonitoring the monitored generated / pre-generated datasets, (2) mcmcEndOfStudy: For each waitWidth and alertK, the average sample size, bias, mse, and errors plus the ecdf of sample size and bias, and (3) the function inputs
}
\description{
The Second Generation p-Value (SGPV) incorporates pre-specified clinical context
to monitor studies until either ruling out trivial or impactful effects.  Operating
characteristics may be determined via.  This function allows the user to use either
their own generated data or to generate their own data from any of R's random data
generators.
}
\details{
Data are either provided from one's own simulations or via monte-carlo simulations of data, using \code{dataGeneration},
\code{dataGenArgs}, \code{effectGeneration}, and \code{effectGenArgs}.  When one's own data are provided, it must be in matrix
format with at least the columns: 'theta', 'est', 'lo' (lower bound), and 'up' (upper bound).  Otherwise, data are generated
until the final \code{maxAlertSteps} observations affirm a stopping rule under unrestricted sample size.  Without \code{dataGeneration},
\code{dataGenArgs}, \code{effectGeneration}, and \code{effectGenArgs} no further data may be generated.  If pre-generated data
are insufficient for unrestricted monitoring, the user is notified to generate more data.

Under working knowledge of the data / clinical context, indicate the point null and clinical guideposts (2 guideposts for one-sided
investigations and 4 guideposts for two-sided investigations).  For a one-sided investigation, leave both of the lower (or upper)
deltas (ie \code{deltaL2} \code{deltaL1} or \code{deltaG1} \code{deltaG2}) set to NA.

To reduce false discoveries, the study waits until the interval of interest (currently coded as the Confidence Interval)
achieves an expected maximum width.  The expected width assumes the standard deviation provided in \code{dataGenArgs}.
Assuming larger variability than the truth yields a longer wait time and more conservative error.

Acruing data are monitored by intervals as specified with \code{modelFit} which may be user defined.

Monitoring rules: An alert is raised when the effect, theta, is ruled to be not trivial or not meaningful (using sgpv).
To reduce bias, increase coverage, and improve operating characteristics in general, the study may require the same alert
be raised k observations later.  To investigate the ideal choice of k, user may specify \code{lookSteps} up until
\code{maxAlertSteps}.

While data are simulated to allow unrestricted sample size, practically the study may feasibly allow up to \code{maxN} units
to accrue.  Further, outcomes are simulated to occur simultaneously.  However, there may be a lag time in Outcomes.
\code{lagOutcomeN} indicates the number of units that are yet being followed.  A study may stop with unrestricted sample size
yet may still have a lag on observing additionally accrued units.

Note: User may choose to save the generated or pre-specified data.  This can be substantive in size.  However, when
generating data from a location-shift family, it may be used to fully explore a range of shifted intervention effects (theta).
}
