% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgpvAMrules.R
\name{sgpvAMrules}
\alias{sgpvAMrules}
\title{amSgpvRules}
\usage{
sgpvAMrules(mcmcMonitoring, fork = TRUE, socket = TRUE,
  cores = detectCores(), waitTime, lookSteps, kSteps, maxAlertSteps,
  monitoringIntervalLevel, getUnrestricted, maxN, lagOutcomeN)
}
\description{
Generate mcmc simulations of adaptive monitoring with parallel computing
}
